+++ engine/server/library/serverGame/src/shared/object/ServerObject.h
@@ -550,7 +550,7 @@ public:
 	virtual bool                  isVisibleOnClient              (const Client & client) const = 0;
 	virtual void                  kill                           ();
 
-	void                          performSocial                  (const NetworkId & target, unsigned long socialType, bool animationOk, bool textOk);
+	void                          performSocial                  (const NetworkId & target, uint32_t socialType, bool animationOk, bool textOk);
 	void                          performSocial                  (const MessageQueueSocial & socialMsg);
 
 	void                          performCombatSpam              (const MessageQueueCombatSpam & combatSpam, bool sendToSelf, bool sendToTarget, bool sendToBystanders);
@@ -582,7 +582,7 @@ public:
 	virtual void                  setOwnerId(const NetworkId &id);
 	void                          setSceneIdOnThisAndContents    (const std::string & sceneId);
 	void                          setPlayerControlled            (bool newValue);
-	void                          speakText                      (NetworkId const &target, unsigned long chatType, unsigned long mood, unsigned long flags, Unicode::String const &speech, int language, Unicode::String const &oob);
+	void                          speakText                      (NetworkId const &target, uint32 chatType, uint32 mood, uint32 flags, Unicode::String const &speech, int language, Unicode::String const &oob);
 	virtual void                  speakText                      (MessageQueueSpatialChat const &spatialChat);
 	virtual void                  hearText                       (ServerObject const &source, MessageQueueSpatialChat const &spatialChat, int chatMessageIndex);
 	void                          teleportObject                 (Vector const & position_w, NetworkId const &targetContainer, std::string const &targetCellName, Vector const &position_p, std::string const &scriptCallback, bool forceLoadScreen = false);
@@ -797,7 +797,7 @@ private:
 	Archive::AutoDeltaVariableCallback<uint32, AuthProcessIdCallback, ServerObject> m_authServerProcessId;
 	Archive::AutoDeltaSet<uint32> m_proxyServerProcessIds;
 
-	Archive::AutoDeltaVariable<unsigned long> m_transformSequence;
+	Archive::AutoDeltaVariable<uint32> m_transformSequence;
 
 	Archive::AutoDeltaVariable<int> m_cacheVersion;
 	Archive::AutoDeltaVariable<bool> m_loadContents;
@@ -843,7 +843,7 @@ protected:
 private:
 	
 	Archive::AutoDeltaVector<TriggerVolumeInfo, ServerObject> m_triggerVolumeInfo;
-	Archive::AutoDeltaMap<std::pair<std::pair<unsigned long, uint64>, MessageToId>, MessageToPayload> m_messageTos;
+	Archive::AutoDeltaMap<std::pair<std::pair<uint32, uint64>, MessageToId>, MessageToPayload> m_messageTos;
 
 	Sphere                                 m_worldSphere;
