+++ engine/server/library/serverGame/src/shared/pvp/Pvp.cpp
@@ -839,18 +839,18 @@ int Pvp::calculateGcwImperialScorePercentile(std::string const & gcwCategory, st
 	if (!getGcwScoreCategory(gcwCategory))
 		return 50; // tie
 
-	uint64 totalImperial = 0ull;
+	uint64 totalImperial = 0;
 	{
 		std::map<std::string, std::pair<int64, int64> >::const_iterator const iterImperial = gcwImperialScore.find(gcwCategory);
 		if (iterImperial != gcwImperialScore.end())
-			totalImperial = static_cast<uint64>(std::max(0ll, iterImperial->second.first)) + static_cast<uint64>(std::max(0ll, iterImperial->second.second));
+			totalImperial = static_cast<uint64>(std::max(static_cast<int64>(0), iterImperial->second.first)) + static_cast<uint64>(std::max(static_cast<int64>(0), iterImperial->second.second));
 	}
 
 	uint64 total = totalImperial;
 	{
 		std::map<std::string, std::pair<int64, int64> >::const_iterator const iterRebel = gcwRebelScore.find(gcwCategory);
 		if (iterRebel != gcwRebelScore.end())
-			total += (static_cast<uint64>(std::max(0ll, iterRebel->second.first)) + static_cast<uint64>(std::max(0ll, iterRebel->second.second)));
+			total += (static_cast<uint64>(std::max(static_cast<int64>(0), iterRebel->second.first)) + static_cast<uint64>(std::max(static_cast<int64>(0), iterRebel->second.second)));
 	}
 
 	// if the score is less than 1, treat it as a tie; note that score is scaled by 1000
@@ -867,7 +867,7 @@ int Pvp::calculateGcwImperialScorePercentile(std::string const & gcwCategory, st
 	if (!getGcwScoreCategory(gcwCategory))
 		return 50; // tie
 
-	uint64 totalImperial = 0ull;
+	uint64 totalImperial = 0;
 	{
 		std::map<std::string, std::pair<uint64, uint64> >::const_iterator const iterImperial = gcwImperialScore.find(gcwCategory);
 		if (iterImperial != gcwImperialScore.end())
