+++ engine/shared/library/sharedDebug/src/linux/DebugHelp.h
@@ -12,7 +12,7 @@
 
 // ======================================================================
 
-typedef unsigned long uint32;
+typedef uint32_t uint32;
 
 // ======================================================================
 
@@ -24,7 +24,7 @@ public:
 	static void remove();
 
 	static void getCallStack(uint32 *callStack, int sizeOfCallStack);
-	static bool lookupAddress(uint32 address, char *libName, char *fileName, int fileNameLength, int &line);
+	static bool lookupAddress(uintptr_t address, char *libName, char *fileName, int fileNameLength, int &line);
 };
 
 // ======================================================================
