+++ engine/shared/library/sharedDebug/src/win32/DebugHelp.h
@@ -10,7 +10,7 @@
 
 // ======================================================================
 
-typedef unsigned long uint32;
+typedef uint32_t uint32;
 
 // ======================================================================
 
@@ -25,7 +25,7 @@ public:
 
 	static void getCallStack(uint32 *callStack, int sizeOfCallStack);
 	static void reportCallStack(int const maxStackDepth = 4);
-	static bool lookupAddress(uint32 address, char *libName, char *fileName, int fileNameLength, int &line);
+	static bool lookupAddress(uintptr_t address, char *libName, char *fileName, int fileNameLength, int &line);
 
 	static bool writeMiniDump(char const *miniDumpFileName=0, PEXCEPTION_POINTERS exceptionPointers=0);
 };
