+++ engine/shared/library/sharedMemoryManager/src/shared/MemoryManager.h
@@ -55,7 +55,7 @@ public:
 	static int             getProcessVmSizeKBytes(const int processId = 0);
 #endif
 
-	static DLLEXPORT void *allocate(size_t size, uint32 owner, bool array, bool leakTest);
+	static DLLEXPORT void *allocate(size_t size, uintptr_t owner, bool array, bool leakTest);
 	static DLLEXPORT void  free(void *pointer, bool array);
 	static DLLEXPORT void  own(void *pointer);
 	static void *          reallocate(void *userPointer, size_t newSize);
